% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVXSynth.R
\name{en_sc_solve}
\alias{en_sc_solve}
\title{Solve for Imbens-Doudchenko elastic net synthetic control weights in CVXR}
\usage{
en_sc_solve(y_t, y_c, lambdas = NULL, alpha = 0.5, t = 10, solv = "MOSEK")
}
\arguments{
\item{y_t}{t_0 X 1   matrix of pre-treatment outcomes for treatment units}

\item{y_c}{t_0 X n_0 matrix of pre-treatment outcomes for donor units}

\item{lambdas}{vector of penalty values}

\item{alpha}{scalar mixing value between L1 and L2 regularisation}

\item{t}{number of lambdas to try (when range not manually specified)}

\item{solv}{what solver to use. default is mosek}
}
\value{
vector of weights
}
\description{
Solve for Imbens-Doudchenko elastic net synthetic control weights in CVXR
}
