% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVXSynth.R
\name{pick_lambda}
\alias{pick_lambda}
\title{Choose regularisation parameter that minimises pseudo-treatment prediction error}
\usage{
pick_lambda(j, Y, lambdas, T0)
}
\arguments{
\item{j}{index of pseudo-treatment unit}

\item{Y}{matrix of control outcomes (wide, where each column is a series)}

\item{lambdas}{sequence of lambda parameters}

\item{T0}{number of pre-treatment periods}
}
\value{
value of lambda that minimises prediction error
}
\description{
Choose regularisation parameter that minimises pseudo-treatment prediction error
}
